+++ src/caselawclient/Client.py
@@ -30,7 +30,7 @@ from caselawclient.models.judgments import Judgment
 from caselawclient.models.press_summaries import PressSummary
 from caselawclient.models.utilities import move
 from caselawclient.search_parameters import SearchParameters
-from caselawclient.types import DocumentURIString
+from caselawclient.types import DocumentIdentifierSlug, DocumentIdentifierValue, DocumentURIString
 from caselawclient.xquery_type_dicts import (
     MarkLogicDocumentURIString,
     MarkLogicDocumentVersionURIString,
@@ -1203,11 +1203,13 @@ class MarklogicApiClient:
 
         return results
 
-    def resolve_from_identifier_slug(self, identifier_uri: str, published_only: bool = True) -> IdentifierResolutions:
+    def resolve_from_identifier_slug(
+        self, identifier_slug: DocumentIdentifierSlug, published_only: bool = True
+    ) -> IdentifierResolutions:
         """Given a PUI/EUI url, look up the precomputed slug and return the
         MarkLogic document URIs which match that slug. Multiple returns should be anticipated"""
         vars: query_dicts.ResolveFromIdentifierSlugDict = {
-            "identifier_uri": DocumentURIString(identifier_uri),
+            "identifier_slug": identifier_slug,
             "published_only": int(published_only),
         }
         raw_results: list[str] = get_multipart_strings_from_marklogic_response(
@@ -1218,14 +1220,8 @@ class MarklogicApiClient:
         )
         return IdentifierResolutions.from_marklogic_output(raw_results)
 
-    def resolve_from_identifier(self, identifier_uri: str, published_only: bool = True) -> IdentifierResolutions:
-        warnings.warn(
-            "resolve_from_identifier deprecated, use resolve_from_identifier_slug instead", DeprecationWarning
-        )
-        return self.resolve_from_identifier_slug(identifier_uri, published_only)
-
     def resolve_from_identifier_value(
-        self, identifier_value: str, published_only: bool = True
+        self, identifier_value: DocumentIdentifierValue, published_only: bool = True
     ) -> IdentifierResolutions:
         """Given a PUI/EUI url, look up the precomputed slug and return the
         MarkLogic document URIs which match that slug. Multiple returns should be anticipated"""